+++ tudou-common/src/main/java/com/tudou/common/base/BaseService.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface BaseService<Record, Example> {
 
-	int countByExample(Example example);
+	long countByExample(Example example);
 
 	int deleteByExample(Example example);
