+++ tudou-upms/tudou-upms-service/src/main/java/com/tudou/upms/server/controller/manage/UpmsUserController.java
@@ -190,7 +190,7 @@ public class UpmsUserController extends BaseController {
 		int pagec = userValid.getPageCurrent();
 		int pages = userValid.getPageSize();
 		List<UpmsUser> rows = upmsUserService.selectByExampleForOffsetPage(upmsUserExample, pagec, pages);
-		int total = upmsUserService.countByExample(upmsUserExample);
+		long total = upmsUserService.countByExample(upmsUserExample);
 		return new UpmsResult(UpmsResultConstant.SUCCESS,rows,pages,pagec,total);
 	}
